♻ refactor(ui): cria `TodosList`
