+++ ui/src/components/TodosHeader.tsx
@@ -1,6 +1,4 @@
-import {
-  Button, HStack, Input, InputGroup, InputRightElement, Select,
-} from '@chakra-ui/react';
+import { Button, HStack, Input, Select } from '@chakra-ui/react';
 import { useContext, useState } from 'react';
 import { TodoField } from '../@types/custom';
 import {
