+++ ui/src/components/TodosList.tsx
@@ -1,25 +1,11 @@
-import { DeleteIcon } from '@chakra-ui/icons';
-import {
-  Badge, Button, Editable, EditableInput, EditablePreview,
-  Flex, HStack, IconButton, Input, Spacer, Text,
-} from '@chakra-ui/react';
+import { Flex, Text } from '@chakra-ui/react';
 import { useContext } from 'react';
-import { Todo } from '../@types/custom';
-import { TodosContext, TodosContextType, TODO_STATUSES } from '../context/TodosContext';
+import { TodosContext, TodosContextType } from '../context/TodosContext';
 import TEST_IDS from '../testIds';
+import TodosItem from './TodosItem';
 
 export default function TodoList() {
-  const {
-    todos, updateTodoStatus, deleteTodo, updateTodo,
-  } = useContext(TodosContext) as TodosContextType;
-
-  const cycleTodoStatus = (todoId: Todo['_id'], currentStatus: Todo['status']) => {
-    const currentIndex = TODO_STATUSES.indexOf(currentStatus);
-    const nextIndex = TODO_STATUSES.length === (currentIndex + 1) ? 0 : currentIndex + 1;
-    const newStatus = TODO_STATUSES[nextIndex] as Todo['status'];
-
-    updateTodoStatus(todoId, newStatus);
-  };
+  const { todos } = useContext(TodosContext) as TodosContextType;
 
   if (!todos || !todos.length) {
     return (
@@ -44,38 +30,7 @@ export default function TodoList() {
       borderRadius="lg"
       p="2"
     >
-      { todos.map(({ _id, description, status, createdAt }) => (
-        <HStack
-          key={ _id }
-          justifyContent="space-between"
-          w="100%"
-          p="1"
-        >
-          <Badge>{new Date(createdAt).toLocaleDateString()}</Badge>
-          <Editable defaultValue={ description }>
-            <EditablePreview />
-            <Input
-              as={ EditableInput }
-              value={ description }
-              onBlur={ (e) => updateTodo(_id, { description: e.target.value }) }
-            />
-          </Editable>
-          <Spacer />
-          <Button
-            size="xs"
-            onClick={ () => cycleTodoStatus(_id, status) }
-          >
-            {status}
-          </Button>
-          <IconButton
-            data-testid={ `${TEST_IDS.todoDelButton}-${_id}` }
-            size="xs"
-            aria-label={ `deletar tarefa ${description}` }
-            icon={ <DeleteIcon /> }
-            onClick={ () => deleteTodo(_id) }
-          />
-        </HStack>
-      ))}
+      { todos.map((todo) => (<TodosItem key={ todo._id } todo={ todo } />))}
     </Flex>
   );
 }
